Split agent tool schemas for chat agent migration
