+++ lib/ai/tools/get-terminal-files.ts
@@ -1,34 +1,17 @@
 import { tool } from "ai";
-import { z } from "zod";
 import type { ToolContext } from "@/types";
 import { uploadSandboxFileToConvex } from "./utils/sandbox-file-uploader";
-import { toolBriefSchema } from "./tool-brief";
 import {
   getSandboxWithFallbackGuard,
   resolveToolErrorMessage,
 } from "./utils/sandbox-fallback";
+import { getTerminalFilesTool } from "./schemas";
 
 export const createGetTerminalFiles = (context: ToolContext) => {
   const { sandboxManager, backgroundProcessTracker } = context;
 
   return tool({
-    description: `Share files from the terminal sandbox with the user as downloadable attachments.
-    
-Usage:
-- Use this tool when the user requests files or needs to download results from the sandbox
-- Provide full file paths (e.g., /home/user/output.txt, /home/user/scan-results.xml)
-- Files are automatically uploaded and made available for download
-- Files larger than 250 MB cannot be shared; reduce, split, or exclude bulky generated/dependency directories before sharing
-- Use this after generating reports, saving scan results, or creating any files the user needs to access
-- Multiple files can be shared in a single call`,
-    inputSchema: z.object({
-      brief: toolBriefSchema,
-      files: z
-        .array(z.string())
-        .describe(
-          "Array of file paths to provide as attachments to the user. Use full paths like /home/user/output.txt",
-        ),
-    }),
+    ...getTerminalFilesTool,
     execute: async ({ files }: { files: string[] }) => {
       try {
         const { sandbox } = await getSandboxWithFallbackGuard({
